优化代码：建立命令类，建立工具类
